requirements.txt
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
docopt==0.6.2
h11==0.12.0
idna==3.3
numpy==1.20.1
outcome==1.1.0
pandas==1.2.3
Pillow==8.4.0
pipreqs==0.4.11
pycparser==2.20
pyOpenSSL==21.0.0
python-dateutil==2.8.1
python-dotenv==0.20.0
pytz==2021.1
requests==2.26.0
selenium==3.6.0
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
webcolors==1.11.1
wsproto==1.0.0
yarg==0.1.9
